bonus route to delete thoughts along with user completed

+++ controllers/userController.js
@@ -49,28 +49,23 @@ module.exports = {
   // Delete a user and remove their thoughts
   async deleteUser(req, res) {
     try {
-      const user = await User.findOneAndRemove({ _id: req.params.userId });
-
+      const userId = req.params.userId;
+  
+      const user = await User.findOneAndRemove({ _id: userId });
+  
       if (!user) {
         return res.status(404).json({ message: 'No such user exists' });
       }
-
-      const thoughtsDeleted = await Thought.deleteMany(
-        { users: req.params.userId }
-      );
-
-      if (!thoughtsDeleted) {
-        return res.status(404).json({
-          message: 'User deleted, but no thoughts found',
-        });
-      }
-
+  
+      const thoughtsDeleted = await Thought.deleteMany({ users: userId });
+  
       res.json({ message: 'User and associated thoughts successfully deleted' });
     } catch (err) {
       console.log(err);
       res.status(500).json(err);
     }
-  },
+  }
+  ,
 
   async updateUser(req, res) {
     try {
